put things where they go
